requirements.txt
tqdm
PyYAML
pandas
minet==0.66.2
trafilatura
beautifulsoup4
requests